@@ -1,4 +1,4 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
 # Pick the frameworks you want:
 require "active_model/railtie"
@@ -21,7 +21,8 @@ module GovernmentFrontend
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     # config.time_zone = 'Central Time (US & Canada)'
 
-    # Explicitly set default locale
+    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
+    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     config.i18n.default_locale = :en
 
     # Explicitly set available locales
@@ -41,5 +42,10 @@ module GovernmentFrontend
 
     # Path within public/ where assets are compiled to
     config.assets.prefix = '/government-frontend'
+
+    # Do not swallow errors in after_commit/after_rollback callbacks.
+    # config.active_record.raise_in_transactional_callbacks = true
+
+    GdsApi.config.always_raise_for_not_found = true
   end
 end
