@@ -59,7 +59,10 @@ class DetailedGuideTest < ActionDispatch::IntegrationTest
   test 'detailed guide that only applies to a set of nations, with alternative urls' do
     setup_and_visit_content_item('national_applicability_alternative_url_detailed_guide')
 
-    assert_has_component_metadata_pair('Applies to', 'England, Scotland, and Wales (see guidance for <a href="http://www.dardni.gov.uk/news-dard-pa022-a-13-new-procedure-for" rel="external">Northern Ireland</a>)')
+    assert_has_component_metadata_pair(
+      'Applies to',
+      'England, Scotland, and Wales (see guidance for <a rel="external" href="http://www.dardni.gov.uk/news-dard-pa022-a-13-new-procedure-for">Northern Ireland</a>)'
+    )
   end
 
   test "translated detailed guide" do
